4209: Make WebView reusable
